refactor(ontology): delete unused components and services
