+++ src/app/project/ontology/ontology-form/ontology-form.component.ts
@@ -12,7 +12,7 @@ import {
 import { DspApiConnectionToken, existingNamesValidator } from '@dasch-swiss/dsp-ui';
 import { CacheService } from 'src/app/main/cache/cache.service';
 import { ErrorHandlerService } from 'src/app/main/error/error-handler.service';
-import { ResourceClassFormService } from '../resource-class-form/resource-class-form.service';
+import { OntologyService } from '../ontology.service';
 
 export interface NewOntology {
     projectIri: string;
@@ -93,7 +93,7 @@ export class OntologyFormComponent implements OnInit {
         private _cache: CacheService,
         private _errorHandler: ErrorHandlerService,
         private _fb: FormBuilder,
-        private _resourceClassFormService: ResourceClassFormService
+        private _ontologyService: OntologyService
     ) { }
 
     ngOnInit() {
@@ -119,7 +119,7 @@ export class OntologyFormComponent implements OnInit {
                 (response: ReadOntology) => {
                     // add values to the ontology form
                     this.ontologyForm.controls['name'].disable();
-                    const name = this._resourceClassFormService.getOntologyName(this.iri);
+                    const name = this._ontologyService.getOntologyName(this.iri);
                     this.ontologyForm.controls['name'].setValue(name);
                     this.ontologyForm.controls['label'].setValue(response.label);
                     this.ontologyForm.controls['label'].setValidators(
